+++ explorer/models/getImageLabel.ts
@@ -8,7 +8,6 @@ import { ImageDesc } from 'dockerode';
 import * as moment from 'moment';
 import * as vscode from 'vscode';
 import { extractRegExGroups } from '../../src/utils/extractRegExGroups';
-import { trimWithElipsis } from '../utils/utils';
 
 // If options not specified, retrieves them from user settings
 export function getImageLabel(fullTag: string, image: ImageDesc, labelTemplate: string, options?: { truncateLongRegistryPaths: boolean, truncateMaxLength: number }): string {
@@ -30,6 +29,20 @@ export function getImageLabel(fullTag: string, image: ImageDesc, labelTemplate: 
     assert(!label.match(/{|}/), "Unreplaced token");
 
     return label;
+}
+
+export function trimWithElipsis(str: string, max: number = 10): string {
+    const elipsis: string = "...";
+    const len: number = str.length;
+
+    if (max <= 0 || max >= 100) { return str; }
+    if (str.length <= max) { return str; }
+    if (max < 3) { return str.substr(0, max); }
+
+    const front: string = str.substr(0, (len / 2) - (-0.5 * (max - len - 3)));
+    const back: string = str.substr(len - (len / 2) + (-0.5 * (max - len - 3)));
+
+    return front + elipsis + back;
 }
 
 /**
