+++ extension.bundle.ts
@@ -23,8 +23,7 @@ export { ext } from './src/extensionVariables';
 export { nonNullProp } from './src/utils/nonNull';
 export { IKeytar } from './src/utils/keytar';
 export { throwDockerConnectionError, internal } from './explorer/utils/dockerConnectionError';
-export { getImageLabel } from './explorer/models/getImageLabel';
-export { trimWithElipsis } from './explorer/utils/utils';
+export { getImageLabel, trimWithElipsis } from './explorer/models/getImageLabel';
 export { isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from "./src/utils/osVersion";
 export { LineSplitter } from './src/debugging/coreclr/lineSplitter';
 export { CommandLineBuilder } from './src/debugging/coreclr/commandLineBuilder';
