+++ src/commands/azure/acr-log-utils/style/fabric-components/css/vscmdl2-icons.css
