+++ src/commands/azure/acr-log-utils/style/fabric-components/fonts/vscmdl2-icons-d3699964.woff
