+++ src/commands/azure/acr-log-utils/style/fabric-components/fonts/vscmdl2-icons.ttf
