+++ src/commands/azure/acr-log-utils/style/fabric-components/microsoft-ui-fabric-assets-license.pdf
