+++ src/commands/azure/acr-log-utils/style/stylesheet.css
