+++ src/commands/build-image.ts
@@ -6,8 +6,8 @@
 import * as path from "path";
 import * as vscode from "vscode";
 import { IActionContext } from "vscode-azureextensionui";
-import { delay } from "../../explorer/utils/utils";
 import { ext } from "../extensionVariables";
+import { delay } from "../utils/delay";
 import { quickPickDockerFileItem } from "../utils/quick-pick-file";
 import { quickPickWorkspaceFolder } from "../utils/quickPickWorkspaceFolder";
 import { addImageTaggingTelemetry, getTagFromUserInput } from "./tag-image";
