+++ src/commands/tag-image.ts
@@ -7,9 +7,9 @@ import vscode = require('vscode');
 import { IActionContext } from 'vscode-azureextensionui';
 import { ImageNode } from '../../explorer/models/imageNode';
 import { RootNode } from '../../explorer/models/rootNode';
-import { delay } from '../../explorer/utils/utils';
 import { configurationKeys } from '../constants';
 import { ext } from '../extensionVariables';
+import { delay } from '../utils/delay';
 import { docker } from '../utils/docker-endpoint';
 import { extractRegExGroups } from '../utils/extractRegExGroups';
 import { ImageItem, quickPickImage } from '../utils/quick-pick-image';
