+++ src/utils/azureUtilityManager.ts
@@ -11,11 +11,11 @@ import { Subscription } from 'azure-arm-resource/lib/subscription/models';
 import { ServiceClientCredentials } from 'ms-rest';
 import * as vscode from 'vscode';
 import { addExtensionUserAgent, callWithTelemetryAndErrorHandling, IActionContext, parseError, UserCancelledError } from 'vscode-azureextensionui';
-import { openExternal } from '../../explorer/utils/openExternal';
 import { AzureAccount, AzureSession } from '../../typings/azure-account.api';
 import { MAX_CONCURRENT_SUBSCRIPTON_REQUESTS } from '../constants';
 import { AsyncPool } from './asyncpool';
 import { getSubscriptionId, getTenantId } from './nonNull';
+import { openExternal } from './openExternal';
 
 /* Singleton for facilitating communication with Azure account services by providing extended shared
   functionality and extension wide access to azureAccount. Tool for internal use.
