+++ src/utils/delay.ts
@@ -3,20 +3,6 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-export function trimWithElipsis(str: string, max: number = 10): string {
-    const elipsis: string = "...";
-    const len: number = str.length;
-
-    if (max <= 0 || max >= 100) { return str; }
-    if (str.length <= max) { return str; }
-    if (max < 3) { return str.substr(0, max); }
-
-    const front: string = str.substr(0, (len / 2) - (-0.5 * (max - len - 3)));
-    const back: string = str.substr(len - (len / 2) + (-0.5 * (max - len - 3)));
-
-    return front + elipsis + back;
-}
-
 export async function delay(ms: number): Promise<void> {
     return new Promise<void>(resolve => {
         setTimeout(() => { resolve(); }, ms);
