+++ src/utils/openExternal.ts
