+++ src/utils/quick-pick-image.ts
@@ -8,9 +8,9 @@ import * as path from "path";
 import vscode = require('vscode');
 import { IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
 import { throwDockerConnectionError } from '../../explorer/utils/dockerConnectionError';
-import { delay } from '../../explorer/utils/utils';
 import { addImageTaggingTelemetry, getTagFromUserInput } from '../commands/tag-image';
 import { ext } from '../extensionVariables';
+import { delay } from './delay';
 import { docker } from './docker-endpoint';
 import { Item } from './quick-pick-file';
